-Updated all try..catch error messages adding the stack trace of the exception -Added TibiaTek Multi-Client Patcher

+++ trunk/TibiaTekBot2/DatReader.vb
@@ -50,7 +50,7 @@ Public Module DatReaderModule
             End Get
         End Property
 
-        Public Function GetInfo(ByVal ItemID As UShort) As DatTile
+        Public Function GetInfo(ByVal ItemID As Integer) As DatTile
             If ItemID >= DatTiles.Length Then Return Nothing
             Return DatTiles(ItemID)
         End Function
